requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
Flask-Migrate>=4.0
Flask-Bcrypt>=1.0
Flask-JWT-Extended>=4.6
python-dotenv>=1.0
